pyproject.toml
[tool.poetry]
name = "recruitment-system-ai"
version = "0.1.0"
description = ""
authors = ["Zimaw AI Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
loguru = "^0.7.3"
pydantic = {extras = ["email"], version = "^2.10.4"}
torchvision = {version = "^0.20.1+cpu", source = "pytorch"}
torch = {version = "^2.5.1+cpu", source = "pytorch"}
docling = "^2.14.0"
openai = "^1.58.1"
orjson = "^3.10.12"
black = "^24.10.0"
langdetect = "^1.0.9"
unstructured = {extras = ["pdf"], version = "^0.16.11"}
pdfplumber = "^0.11.4"
unidecode = "^1.3.8"
pypdf2 = "^3.0.1"
fuzzywuzzy = "^0.18.0"
groq = "^0.13.1"
google-generativeai = "^0.8.3"
python-levenshtein = "^0.26.1"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
